backend/requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# OpenAI Integration
langchain-openai
openai

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# Vector Store
faiss-cpu

# YouTube
youtube-transcript-api